+++ src/app/(auth)/login/login-form.tsx
@@ -11,7 +11,7 @@ import { Button } from '@/components/ui/button';
 import { Icons } from '@/components/icon/icons';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { LoginInput, loginSchema } from '@/lib/validations/schemas';
-import { useAuth } from '../../hooks/use-auth';
+import { useAuth } from '../../../hooks/use-auth';
 import LoaderCircleIcon from '@/components/shared/loader-circle';
 
 type LoginTabValue = 'admin' | 'staff';
